[refactor] confirmModal 파일 위치를 common폴더로 이동
